Fix #217
Update umap_.transform to pass a copy of self.embedding_ to
optimize_layout

Fix #179 converted embeddings to np.float32 in the call to
optimize_layout, but the copy argument of np.ndarray.astype was set to
False. However, this had the effect of copying embedding_ when the dtype
was not np.float32 and passing the actual embedding_ data when it was
already np.float32. Given that optimize_layout should only be "touching"
the embedding_ but not modifying it, using copy=True in
np.ndarray.astype seemed the approriate fix.

This commit also adds a unit test to test_umap.py to handle these cases.