distinction of cross entropy and KL divergence
I'm not entirely sure of the inner workings of the algorithm, but when reading this documentation and comparing with other sources I found that the expression for what was named 'cross entropy' did not seem correct. Instead, there are two separate terms describing two separate KL divergences (one for the change in entropy in the probability of the simplex existing, and one for not existing). It is not clear in the text (even with my suggestions) why one needs to divergences. I make no claims to the workings of the algorithm, but only suggest changes to the descriptions of the mathematics in the documentation. 

I hope it makes sense, and thanks to everyone for putting this great resource together!