Added the ability to prune distance greater than a given threshold from the UMAP-graph before embedding.  This is particularly useful if you have a bounded distance and points that are maximally distant from all other points.  In this case UMAP will connect to a random set of points at a transformed distance of 0 due to our connectivity axiom.
Added a default DISCONNECTION_DISTANCE for known bounded distances.  The dictionary will need to be updated as new bounded distances are added.
Added the ability to pass in a value for disconnection_distance.  This defaults to the maximal distance for bounded distances.
Exposed an _inverse map to allow user to make use of umap internal objects even when they've called UMAP with unique=True.
Cut edges greater than or equal to the disconnection_distance for the sparse UMAP case.
Fixed a bug in spectral.py where kwds weren't being passed SPECIAL_METRICS.