Nose removed, Skip test if not TF and improved tests
Nose dependency is completely removed from all tests, and replaced with corresponding pytest analogous statements/constructs.

The main and most important change to this test module concerns the integration of `pytest.mark.skipif` decorator to skip all the tests requiring the use of TensorFlow.
Since `tensorflow` is an optional dependency for umap (and so `ParametricUMAP`), tests will be excluded if requirements are not satisfied.
This may be useful in cases `tensorflow` requirements do not match with the environment umap is installed into (e.g. Python version), still allowing other tests to run.