First implementation to Fix Issue #590
This commit includes a new implementation for the `chunked_parallel_spatial_metric` in `distances` to fix issue #590 causing Python 3.8 to fail tests in the case when `X` and `Y` parameters are passed.

In more details, this new implementation addresses the issue and also provides a more compact version to avoid code duplications on the two branches (i.e. `Y is|is not None`).

In order to tests the two implementations, a special fixture/decorator is created to compare calculatations with a quick-fix implementation of the previous original `chunked_parallel_special_metric` as implemented by Leland.