Allow points to be passed to umap.plot.points
The actual points to be plotted are usually
taken from the reducer returned by umap.UMAP.
However it is possible to generate other sets
of points by subsequently calling reducer.transform().

This change allows the plot.points call to
optionally accept the points array so that
other sets of points can easily be plotted
other than those that defined the reducer.