Drop requirements.txt, embrace setup.py for deps
File requirements.txt duplicates the UMAP dependency information that we
must put in setup.py. It makes sense then to drop it, and to trust the
installation of the UMAP package as sufficient for setting up its own
dependencies. The README is updated in this sense, and the CI pipelines
are streamlined.