Import pynndescent lazily to speed up import.
Importing umap tended to take long; pynndescent is the culprit. This
moves the substantial cost of importing pynndescent from import time
to the first use. Among other things, this is likely to make it much
nicer to run `script.py --help` and other operations that actually do
not use UMAP on programs that import UMAP.